Introduce Key/KeyRing wrappers
The goal of these classes is to greatly reduce the needed static helper
classes that work on BouncyCastle key ring and key objects.
Ideally only the pgp package uses BouncyCastle code, the rest of the
app should use this abstraction.

PublicKeyRing and SecretKeyRing exist only for cases where it is vital
that a specific kind of key ring is needed, for instance when it comes
to uploading to a key server, where a secret key ring must never be
used by accident.