Fix Key serialization
The data returned by key.getEncoded() cannot be decoded on its
own by PGPObjectFactory. That'd require a key ring and with that
a master key.

The new serialization serializes the Key object and allows restoring
it accurately without creating a key ring as a workaround.

In PgpKeyOperation this requires adjusting the public key, as that
one is queried with isMasterKey(), and it is considered a master key
if subSigs == null. To avoid this, subSigs is initialized with an
empty ArrayList, resulting in a true subkey.